+++ packages/react-native-reanimated/src/css/platform/native/normalization/common/index.ts
@@ -4,4 +4,3 @@ export {
   normalizeDuration,
   normalizeTimingFunction,
 } from './settings';
-export * from './shorthand';
